support sd map in issueCredentialFromJson
